+++ NEWS.md
@@ -1,7 +1,9 @@
-# mizer 2.5.1.9004
+# mizer 2.5.1.9005
 
 - `addSpecies()` now correctly preserves all `species_params` of the existing
   model.
+- `addSpecies()` no longer requires new species to grow to maximum size, only
+  maturity size is required.
 - New species parameter `w_repro_max` giving the size at which a species 
   invests 100% of its energy into reproduction. Set to `w_max` by default.
 - `getDiet()` now also includes the contribution of the external encounter rate